requirements.txt
Django==1.11.9
PyMySQL==0.8.0
Pillow==5.2.0
